fix difftree hashing logic
this change makes it so that the hash of a node only depends on its
children and not also its next sibling. this correctly implements the
idea of nodes with equal hashes having the same subtree.

this change also makes it so that the key value of the AX_NODE container
also includes its index in the childlist of its parent, this allows the
full key of different children under the same parent to be
differentiated.